Finalise image exporter ready for review
