+++ Tools/dea_tools/app/imageexport.py
@@ -1,389 +1,787 @@
-# notebookapp_imageexport.py
-'''
-This file contains functions for creating an interactive map for 
-selecting satellite imagery and exporting image files.
+# -*- coding: utf-8 -*-
+"""
+Image export widget, which can be used to interactively select and 
+export satellite imagery from multiple DEA products.
+"""
 
-Available functions:
-    select_region_app
-    export_image_app
-
-Last modified: September 2021
-'''
-
-# Load modules
+# Import required packages
+import fiona
+import sys
 import datacube
+import warnings
+import matplotlib.pyplot as plt
+from datacube.utils.geometry import CRS
+from ipyleaflet import (
+    WMSLayer,
+    basemaps,
+    basemap_to_tiles,
+    Map,
+    DrawControl,
+    WidgetControl,
+    LayerGroup,
+    LayersControl,
+    GeoData,
+)
+from traitlets import Unicode
+from ipywidgets import (
+    GridspecLayout,
+    Button,
+    Layout,
+    HBox,
+    VBox,
+    HTML,
+    Output,
+)
+import json
 import itertools
 import numpy as np
-import matplotlib.pyplot as plt
-from odc.ui import select_on_a_map
-from datacube.utils.geometry import CRS
-from datacube.utils import masking
+import geopandas as gpd
+from io import BytesIO
+import ipywidgets as widgets
+import datetime
 from skimage import exposure
-from ipyleaflet import (WMSLayer, basemaps, basemap_to_tiles)
-from traitlets import Unicode
+from skimage.filters import unsharp_mask
 
-import sys
-
-sys.path.insert(1, '../Tools/')
-from dea_tools.spatial import reverse_geocode
+from datacube.utils import masking
+from datacube.utils.geometry import Geometry
+import dea_tools.app.widgetconstructors as deawidgets
 from dea_tools.dask import create_local_dask_cluster
+from dea_tools.spatial import reverse_geocode
 from dea_tools.datahandling import pan_sharpen_brovey
 
 
-def select_region_app(date,
-                      satellites,
-                      size_limit=10000):
-    '''
-    An interactive app that allows the user to select a region from a
-    map using imagery from Sentinel-2 and Landsat. The output of this
-    function is used as the input to `export_image_app` to export high-
-    resolution satellite images.
-    
-    Last modified: September 2021
-
-    Parameters
-    ----------
-    date : str
-        The exact date used to plot imagery on the interactive map
-        (e.g. `date='1988-01-01'`). 
-    satellites : str
-        The satellite data to plot on the interactive map. Three options
-        are currently supported:
-            "Landsat": data from the Landsat 5, 7 and 8 satellites
-            "Sentinel-2": data from Sentinel-2A and Sentinel-2B
-            "Sentinel-2 NRT": most recent 'near real time' data from 
-            Sentinel-2A and Sentinel-2B (use this to obtain imagery 
-            acquired in the past three months).
-    size_limit : int, optional
-        An optional size limit for the area selection in sq km.
-        Defaults to 10000 sq km.
-        
-    Returns
-    -------
-    A dictionary containing 'geopolygon' (defining the area to 
-    export imagery from), 'date' (date used to export imagery), and 
-    'satellites' (the satellites from which to extract imagery). These
-    are passed to the `export_image_app` function to export the image.
-    '''
-
-    ########################
-    # Select and load data #
-    ########################
-
-    # Load DEA WMS
-    class TimeWMSLayer(WMSLayer):
-        time = Unicode('').tag(sync=True, o=True)
-    
-    # WMS layers
-    wms_params = {'Landsat': 'ga_ls_ard_3',
-                  'Sentinel-2': 's2_ard_granule_nbar_t',
-                  'Sentinel-2 NRT': 's2_nrt_granule_nbar_t'}
-
-    time_wms = TimeWMSLayer(url='https://ows.dea.ga.gov.au/',
-                            layers=wms_params[satellites],
-                            time=date,
-                            format='image/png',
-                            transparent=True,
-                            attribution='Digital Earth Australia')
-
-    # Plot interactive map to select area
-    basemap = basemap_to_tiles(basemaps.OpenStreetMap.Mapnik)
-    geopolygon = select_on_a_map(height='600px',
-                                 layers=(
-                                     basemap,
-                                     time_wms,
-                                 ),
-                                 center=(-25.18, 134.18),
-                                 zoom=4)
-
-    # Test size of selected area
-    area = geopolygon.to_crs(crs=CRS('epsg:3577')).area / 1000000
-    if area > size_limit:
-        print(f'Warning: Your selected area is {area:.00f} sq km. '
-              f'Please select an area of less than {size_limit} sq km.'
-              f'\nTo select a smaller area, re-run the cell '
-              f'above and draw a new polygon.')
-
-    else:
-        return {'geopolygon': geopolygon, 
-                'date': date,
-                'satellites': satellites}
-      
-
-def export_image_app(geopolygon,
-                     date,
-                     satellites,
-                     style='True colour',
-                     pansharpen=True,
-                     resolution=None,
-                     vmin=0,
-                     vmax=2000,
-                     percentile_stretch=None,
-                     power=None,
-                     image_proc_funcs=None,
-                     output_format="jpg",
-                     standardise_name=False):
-    '''
-    Exports Digital Earth Australia satellite data as an image file
-    based on the extent and time period selected using 
-    `select_region_app.` The function supports Sentinel-2 and Landsat 
-    data, creating True and False colour images, and applying 
-    pansharpening to increase the resolution of Landsat 7 and 8 imagery.
-        
-    By default, files are named to match the DEA Imagery and Animations 
-    folder naming convention:
-    
-        "<product> - <YYYY-MM-DD> - <site, state> - <description>.png" 
-        
-    Set `standardise_name=True` for a machine-readable name:
-    
-        "<product>_<YYYY-MM-DD>_<site-state>_<description>.png" 
-    
-    Last modified: September 2021
-
-    Parameters
-    ----------
-    geopolygon : datacube.utils.geometry object
-        A datacube geopolygon providing the spatial bounds used to load
-        satellite data.
-    date : str
-        The exact date used to extract imagery 
-        (e.g. `date='1988-01-01'`). 
-    satellites : str
-        The satellite data to be used to extract imagery. Three 
-        options are currently supported:
-            "Landsat": data from the Landsat 5, 7 and 8 satellites
-            "Sentinel-2": data from Sentinel-2A and Sentinel-2B
-            "Sentinel-2 NRT": most recent 'near real time' data from 
-            Sentinel-2A and Sentinel-2B (use this to obtain imagery 
-            acquired in the past three months).
-    style : str, optional
-        The style used to produce the image. Two options are currently
-        supported:
-            "True colour": Creates a true colour image using the red, 
-            green and blue satellite bands
-            "False colour": Creates a false colour image using 
-            short-wave infrared, infrared and green satellite bands.
-            The specific bands used vary between Landsat and Sentinel-2.  
-    pansharpen: bool, optional
-        Whether to apply pansharpening (using the Brovey Transform) to 
-        increase the resolution of Landsat imagery from 30 m to 15 m 
-        pixels. This will only be applied if all the following are 
-        true: 
-            - Landsat 7 or 8 data (with panchromatic band)
-            - `style` is set to "True colour"
-            - `resolution` is set to `None` (see below).
-    resolution : tuple, optional
-        The spatial resolution to load data. By default, the tool will 
-        automatically set the best possible resolution depending on the 
-        satellites selected (i.e 30 m for Landsat, 10 m for Sentinel-2). 
-        Increasing this (e.g. to resolution = (-100, 100)) can be useful
-        for loading large spatial extents.
-    vmin, vmax : int or float
-        The minimum and maximum surface reflectance values used to 
-        clip the resulting imagery to enhance contrast. 
-    percentile_stretch : tuple of floats, optional
-        An tuple of two floats (between 0.00 and 1.00) that can be used 
-        to clip the imagery to based on percentiles to get more control 
-        over the brightness and contrast of the image. The default is 
-        None; '(0.02, 0.98)' is equivelent to `robust=True`. If this 
-        parameter is used, `vmin` and `vmax` will have no effect.
-    power : float, optional
-        Raises imagery by a power to reduce bright features and 
-        enhance dark features. This can add extra definition over areas
-        with extremely bright features like snow, beaches or salt pans.
-    image_proc_funcs : list of funcs, optional
-        An optional list containing functions that will be applied to 
-        the output image. This can include image processing functions 
-        such as increasing contrast, unsharp masking, saturation etc. 
-        The function should take AND return a `numpy.ndarray` with 
-        shape [y, x, bands]. If your function has parameters, you 
-        can pass in custom values using a lambda function, e.g.:
-        `image_proc_funcs = [lambda x: skimage.filters.unsharp_mask(x, radius=5, amount=0.2)]`
-    output_format : str, optional
-        The output file format of the image. Valid options include "jpg"
-        and "png". Defaults to "jpg".
-    standardise_name : bool, optional
-        Whether to export the image file with a machine-readable 
-        file name (e.g. `<product>_<YYYY-MM-DD>_<site-state>_<description>.png`)        
-    '''
-    
-    ###########################
-    # Set up satellite params #
-    ###########################
-
-    sat_params = {
-        'Landsat': {
-            'products': ['ga_ls5t_ard_3', 'ga_ls7e_ard_3', 'ga_ls8c_ard_3'],
-            'resolution': [-30, 30],
-            'styles': {
-                'True colour': ['nbart_red', 'nbart_green', 'nbart_blue'],
-                'False colour': ['nbart_swir_1', 'nbart_nir', 'nbart_green']
-            }
-        },
-        'Sentinel-2': {
-            'products': ['s2a_ard_granule', 's2b_ard_granule'],
-            'resolution': [-10, 10],
-            'styles': {
-                'True colour': ['nbart_red', 'nbart_green', 'nbart_blue'],
-                'False colour': ['nbart_swir_2', 'nbart_nir_1', 'nbart_green']
-            }
-        },
-        'Sentinel-2 NRT': {
-            'products': ['s2a_nrt_granule', 's2b_nrt_granule'],
-            'resolution': [-10, 10],
-            'styles': {
-                'True colour': ['nbart_red', 'nbart_green', 'nbart_blue'],
-                'False colour': ['nbart_swir_2', 'nbart_nir_1', 'nbart_green']
-            }
+# WMS params and satellite style bands
+sat_params = {
+    'ga_ls_ard_3': {
+        'products': ['ga_ls5t_ard_3', 'ga_ls7e_ard_3', 'ga_ls8c_ard_3'],
+        'styles': {
+            'True colour':
+                ('true_colour', ['nbart_red', 'nbart_green', 'nbart_blue']),
+            'False colour':
+                ('false_colour', ['nbart_swir_1', 'nbart_nir', 'nbart_green'])
+        }
+    },
+    's2_ard_granule_nbar_t': {
+        'products': ['s2a_ard_granule', 's2b_ard_granule'],
+        'styles': {
+            'True colour':
+                ('simple_rgb', ['nbart_red', 'nbart_green', 'nbart_blue']),
+            'False colour': ('infrared_green',
+                             ['nbart_swir_2', 'nbart_nir_1', 'nbart_green'])
+        }
+    },
+    's2_nrt_granule_nbar_t': {
+        'products': ['s2a_nrt_granule', 's2b_nrt_granule'],
+        'styles': {
+            'True colour':
+                ('simple_rgb', ['nbart_red', 'nbart_green', 'nbart_blue']),
+            'False colour': ('infrared_green',
+                             ['nbart_swir_2', 'nbart_nir_1', 'nbart_green'])
         }
     }
-        
-    #############
-    # Load data #
-    #############
+}
+
+
+def make_box_layout():
+    return Layout(
+        #         border='solid 1px black',
+        margin='0px 10px 10px 0px',
+        padding='5px 5px 5px 5px',
+        width='100%',
+        height='100%',
+    )
+
+
+def create_expanded_button(description, button_style):
+    return Button(
+        description=description,
+        button_style=button_style,
+        layout=Layout(width="auto", height="auto"),
+    )
+
+
+def update_map_layers(self):
+    """
+    Updates map to add new DEA layers, styles or basemap when selected
+    using menu options. Triggers data reload by resetting load params 
+    and output arrays.
+    """
+
+    # Clear data load params to trigger data re-load
+    self.rgb_array = None
+    self.sensor = None
+    self.load_params = None
+    self.query_params = None
+
+    # Clear all layers and add basemap
+    self.map_layers.clear_layers()
+    self.map_layers.add_layer(self.basemap)
+
+    # Get style name for specific satellite sensor
+    style = sat_params[self.dealayer]['styles'][self.style][0]
+
+    # Add DEA layers over the top of the basemap
+    dea_layer = deawidgets.create_dea_wms_layer(self.dealayer,
+                                                self.date,
+                                                styles=style)
+    self.map_layers.add_layer(dea_layer)
+    
+    
+    
+def extract_data(self):
     
     # Connect to datacube database
-    dc = datacube.Datacube(app='Exporting_satellite_images')
+    dc = datacube.Datacube(app="Exporting satellite images")
 
     # Configure local dask cluster
-    create_local_dask_cluster()
+    client = create_local_dask_cluster(return_client=True,
+                                       display_client=True)
+
+    # Convert to geopolygon
+    geopolygon = Geometry(geom=self.gdf_drawn.geometry[0],
+                          crs=self.gdf_drawn.crs)
 
     # Create query after adjusting interval time to UTC by
     # adding a UTC offset of -10 hours. This results issues
     # on the east coast of Australia where satelite overpasses
     # can occur on either side of 24:00 hours UTC
-    start_date = np.datetime64(date) - np.timedelta64(10, 'h')
-    end_date = np.datetime64(date) + np.timedelta64(14, 'h')
-    query_params = {
+    start_date = np.datetime64(self.date) - np.timedelta64(10, 'h')
+    end_date = np.datetime64(self.date) + np.timedelta64(14, 'h')
+    self.query_params = {
         'time': (str(start_date), str(end_date)),
         'geopolygon': geopolygon
     }
 
     # Find matching datasets
     dss = [
-        dc.find_datasets(product=i, **query_params)
-        for i in sat_params[satellites]['products']
+        dc.find_datasets(product=i, **self.query_params)
+        for i in sat_params[self.dealayer]['products']
     ]
     dss = list(itertools.chain.from_iterable(dss))
-
-    # Get CRS
-    crs = str(dss[0].crs)
     
-    # Get sensor (try/except to account for different S2 NRT metadata)
-    try:
-        sensor = dss[0].metadata_doc['properties']['eo:platform'].capitalize()
-    except: 
-        sensor = dss[0].metadata_doc['platform']['code'].capitalize()
-    sensor = sensor[0:-1].replace('_', '-') + sensor[-1].capitalize()
+    # If data is found
+    if len(dss) > 0:
 
-    # Meets pansharpening requirements
-    can_pansharpen = style == 'True colour' and not resolution and sensor in [
-        'Landsat-7', 'Landsat-8'
-    ]
+        # Get CRS
+        crs = str(dss[0].crs)
 
-    # Set up load params
-    if pansharpen and can_pansharpen:
-        load_params = {
-            'output_crs': crs,
-            'resolution': (-15, 15),
-            'align': (7.5, 7.5),
-            'resampling': 'bilinear'
-        }
+        # Get sensor (try/except to account for different S2 NRT metadata)
+        try:
+            sensor = dss[0].metadata_doc['properties'][
+                'eo:platform'].capitalize()
+        except:
+            sensor = dss[0].metadata_doc['platform']['code'].capitalize(
+            )
+        self.sensor = sensor[0:-1].replace(
+            '_', '-') + sensor[-1].capitalize()
 
-        # Add panchromatic to list of true colour bands
-        sat_params[satellites]['styles']['True colour'] += [
-            'nbart_panchromatic'
+        # Meets pansharpening requirements
+        can_pansharpen = self.style == 'True colour' and self.sensor in [
+            'Landsat-7', 'Landsat-8'
         ]
 
+        # Set up load params
+        if self.pansharpen and can_pansharpen:
+            self.load_params = {
+                'measurements':
+                    sat_params[self.dealayer]['styles'][self.style][1] +
+                    ['nbart_panchromatic'],
+                'resolution': (-self.resolution, self.resolution),
+                'align': (7.5, 7.5),
+                'output_crs':
+                    crs
+            }
+
+        else:
+
+            # Use resolution if provided, otherwise use default
+            if self.resolution:
+                sat_params[self.dealayer]['resolution'] = (
+                    -self.resolution, self.resolution)
+
+            self.load_params = {
+                'measurements':
+                    sat_params[self.dealayer]['styles'][self.style][1],
+                'resolution': (-self.resolution, self.resolution),
+                'output_crs':
+                    crs
+            }
+
+        # Load data from datasets
+        print(f'Loading {self.sensor} satellite data')
+        ds = dc.load(datasets=dss,
+                     resampling='bilinear',
+                     group_by='solar_day',
+                     dask_chunks={
+                         'time': 1,
+                         'x': 2048,
+                         'y': 2048
+                     },
+                     **self.load_params,
+                     **self.query_params)
+        ds = masking.mask_invalid_data(ds)
+
+        # Create plain numpy array, optionally after pansharpening
+        if self.pansharpen and can_pansharpen:
+
+            # Perform Brovey pan-sharpening and return three numpy.arrays
+            print(
+                f'Pansharpening {self.sensor} image to 15 m resolution')
+            red_sharpen, green_sharpen, blue_sharpen = pan_sharpen_brovey(
+                band_1=ds.nbart_red,
+                band_2=ds.nbart_green,
+                band_3=ds.nbart_blue,
+                pan_band=ds.nbart_panchromatic)
+            rgb_array = np.vstack(
+                [red_sharpen, green_sharpen, blue_sharpen])
+
+            # Match histogram to original array
+            from skimage.exposure import match_histograms
+            original_array = ds.drop(
+                'nbart_panchromatic').to_array().squeeze('time').values
+            rgb_array = match_histograms(image=rgb_array,
+                                         reference=original_array,
+                                         multichannel=True)
+
+        # If pansharpening is requested but not possible, deactivate 
+        # pansharpening and reset to 30 m resolution
+        elif self.pansharpen and not can_pansharpen:
+            print(
+                '\nUnable to pansharpen; reverting to 30 m resolution'
+            )
+            self.checkbox_pansharpen.value = False
+            self.text_resolution.disabled = False
+            self.text_resolution.value = 30
+            rgb_array = ds.isel(time=0).to_array().values
+
+        else:
+            rgb_array = ds.isel(time=0).to_array().values
+
+        # Transpose numpy array
+        rgb_array = np.transpose(rgb_array, axes=[1, 2, 0])
+    
+    # Else if no data is returned, return None
     else:
+        rgb_array = None
+        
+    # Close down the dask client
+    client.close()
+        
+    return rgb_array
 
-        # Use resolution if provided, otherwise use default
-        if resolution:
-            sat_params[satellites]['resolution'] = resolution
 
-        load_params = {
-            'output_crs': crs,
-            'resolution': sat_params[satellites]['resolution'],
-            'resampling': 'bilinear'
-        }
-
-    # Load data from datasets
-    ds = dc.load(datasets=dss,
-                 measurements=sat_params[satellites]['styles'][style],
-                 group_by='solar_day',
-                 dask_chunks={
-                     'time': 1,
-                     'x': 3000,
-                     'y': 3000
-                 },
-                 **load_params,
-                 **query_params)
-    ds = masking.mask_invalid_data(ds)
-
-    # Create plain numpy array, optionally after pansharpening
-    if pansharpen and can_pansharpen:
-
-        # Perform Brovey pan-sharpening and return three numpy.arrays
-        print(f'Pansharpening {sensor} image to 15 m resolution')
-        red_sharpen, green_sharpen, blue_sharpen = pan_sharpen_brovey(
-            band_1=ds.nbart_red,
-            band_2=ds.nbart_green,
-            band_3=ds.nbart_blue,
-            pan_band=ds.nbart_panchromatic)
-        rgb_array = np.vstack([red_sharpen, green_sharpen, blue_sharpen])
-
-    else:
-        rgb_array = ds.isel(time=0).to_array().values
-
-    ############
-    # Plotting #
-    ############
-
-    # Create unique file name
-    centre_coords = geopolygon.centroid.coords[0][::-1]
-    site = reverse_geocode(coords=centre_coords)
-    fname = (f"{sensor} - {date} - {site} - {style}, "
-             f"{load_params['resolution'][1]} m resolution.{output_format}")
-
-    # Remove spaces and commas if requested
-    if standardise_name:
-        fname = fname.replace(' - ', '_').replace(', ',
-                                                  '-').replace(' ',
-                                                               '-').lower()
-
-    print(
-        f'\nExporting image to {fname}.\nThis may take several minutes to complete...'
-    )
-
-    # Convert to numpy array
-    rgb_array = np.transpose(rgb_array, axes=[1, 2, 0])
+def plot_data(self, fname):
+    
+    # Data to plot
+    to_plot = self.rgb_array
 
     # If percentile stretch is supplied, calculate vmin and vmax
     # from percentiles
-    if percentile_stretch:
-        vmin, vmax = np.nanpercentile(rgb_array, percentile_stretch)
+    if self.percentile_stretch:
+        vmin, vmax = np.nanpercentile(to_plot, self.percentile_stretch)
+    else:
+        vmin, vmax = self.vmin, self.vmax
 
     # Raise by power to dampen bright features and enhance dark.
     # Raise vmin and vmax by same amount to ensure proper stretch
-    if power:
-        rgb_array = rgb_array**power
-        vmin, vmax = vmin**power, vmax**power
-
+    if self.power < 1.0:
+        with self.status_info:
+            print(f'\nApplying power transformation ({self.power})')
+        to_plot = to_plot ** self.power
+        vmin, vmax = vmin ** self.power, vmax ** self.power
+        
     # Rescale/stretch imagery between vmin and vmax
-    rgb_rescaled = exposure.rescale_intensity(rgb_array.astype(float),
-                                              in_range=(vmin, vmax),
-                                              out_range=(0.0, 1.0))
-
-    # Apply image processing funcs
-    if image_proc_funcs:
-        for i, func in enumerate(image_proc_funcs):
-            print(f'Applying custom function {i + 1}')
-            rgb_rescaled = func(rgb_rescaled)
+    to_plot = exposure.rescale_intensity(to_plot.astype(float),
+                                         in_range=(vmin, vmax),
+                                         out_range=(0.0, 1.0))
+    
+    # Unsharp mask
+    if self.unsharp_mask:
+        with self.status_info:
+            print(f'\nApplying unsharp masking with {self.unsharp_mask_radius} '
+                  f'radius and {self.unsharp_mask_amount} amount')
+        to_plot = unsharp_mask(to_plot, 
+                               radius=self.unsharp_mask_radius, 
+                               amount=self.unsharp_mask_amount)
 
     # Plot RGB
-    plt.imshow(rgb_rescaled)
+    fig, ax = plt.subplots(1, 1, figsize=(12, 9))
+    ax.imshow(to_plot)
 
     # Export to file
-    plt.imsave(fname=fname, arr=rgb_rescaled, format=output_format)
-    print('Finished exporting image.')
+    plt.imsave(fname=fname,
+               arr=np.ascontiguousarray(to_plot),
+               format=self.output_format)
+    plt.show()
+    with self.status_info:
+        print(f'\nImage successfully exported to:\n{fname}.')
+
+
+
+class imageexport_app(HBox):
+
+    def __init__(self):
+        super().__init__()
+
+        ######################
+        # INITIAL ATTRIBUTES #
+        ######################
+
+        # Basemap
+        self.basemap_list = [
+            ("Open Street Map",
+             basemap_to_tiles(basemaps.OpenStreetMap.Mapnik)),
+            ("ESRI World Imagery",
+             basemap_to_tiles(basemaps.Esri.WorldImagery)),
+        ]
+        self.basemap = self.basemap_list[0][1]
+
+        # Satellite data
+        self.date = "2021-12-31"
+        self.dealayer_list = [
+            ("Landsat", 'ga_ls_ard_3'),
+            ("Sentinel-2", 's2_ard_granule_nbar_t'),
+            ("Sentinel-2 NRT", 's2_nrt_granule_nbar_t'),
+        ]
+        self.dealayer = self.dealayer_list[0][1]
+
+        # Styles
+        self.styles_list = ["True colour", "False colour"]
+        self.style = self.styles_list[0]
+
+        # Analysis params
+        self.resolution = 30
+        self.pansharpen = False
+        self.standardise_name = False
+        self.vmin = 50
+        self.vmax = 3000
+        self.percentile_stretch = None  # (1, 99)
+        self.power = 1.0
+        self.output_list = [("JPG", 'jpg'), ("PNG", 'png')]
+        self.output_format = self.output_list[0][1]        
+        self.unsharp_mask = False
+        self.unsharp_mask_radius = 20
+        self.unsharp_mask_amount = 0.3
+        self.max_size = False
+
+        # Drawing params
+        self.target = None
+        self.action = None
+        self.gdf_drawn = None
+
+        # Data load params
+        self.rgb_array = None
+        self.sensor = None
+        self.load_params = None
+        self.query_params = None
+
+        ##################
+        # HEADER FOR APP #
+        ##################
+
+        # Create the Header widget
+        header_title_text = "<h3>Digital Earth Australia satellite image export</h3>"
+        instruction_text = ("<p>Select the desired satellite data, imagery "
+                            "date and image style, zoom in until satellite "
+                            "imagery appears on the map, then draw a "
+                            "rectangle to select an area of imagery to "
+                            "export as a high-resolution image file.</p>")
+        self.header = deawidgets.create_html(
+            f"{header_title_text}{instruction_text}")
+        self.header.layout = make_box_layout()
+
+        #####################################
+        # HANDLER FUNCTION FOR DRAW CONTROL #
+        #####################################
+
+        # Define the action to take once something is drawn on the map
+        def update_geojson(target, action, geo_json):
+
+            # Get data from action
+            self.action = action
+
+            # Clear data load params to trigger data re-load
+            self.rgb_array = None
+            self.sensor = None
+            self.load_params = None
+            self.query_params = None
+
+            # Convert data to geopandas
+            json_data = json.dumps(geo_json)
+            binary_data = json_data.encode()
+            io = BytesIO(binary_data)
+            io.seek(0)
+            gdf = gpd.read_file(io)
+            gdf.crs = "EPSG:4326"
+
+            # Convert to Albers and compute area
+            gdf_drawn_albers = gdf.copy().to_crs("EPSG:3577")
+            m2_per_km2 = 10**6
+            area = gdf_drawn_albers.area.values[0] / m2_per_km2
+            polyarea_label = 'Total area of satellite data to extract'
+            polyarea_text = f"<b>{polyarea_label}</b>: {area:.2f} km<sup>2</sup>"
+
+            # Test area size
+            if self.max_size:
+                confirmation_text = ('<span style="color: #33cc33"> '
+                                     '<b>(Overriding maximum size limit; use with caution as may lead to memory issues)</b></span>')
+                self.header.value = header_title_text + instruction_text + polyarea_text + confirmation_text
+                self.gdf_drawn = gdf                
+            elif area <= 10000:
+                confirmation_text = ('<span style="color: #33cc33"> '
+                                     '<b>(Area to extract falls within '
+                                     'recommended limit)</b></span>')
+                self.header.value = header_title_text + instruction_text + polyarea_text + confirmation_text
+                self.gdf_drawn = gdf
+            else:
+                warning_text = ('<span style="color: #ff5050"> '
+                                '<b>(Area to extract is too large, '
+                                'please select an area less than 10000 '
+                                'km<sup>2)</b></span>')
+                self.header.value = header_title_text + instruction_text + polyarea_text + warning_text
+                self.gdf_drawn = None
+
+        ###########################
+        # WIDGETS FOR APP OUTPUTS #
+        ###########################
+
+        self.status_info = Output(layout=make_box_layout())
+        self.output_plot = Output(layout=make_box_layout())
+
+        #########################################
+        # MAP WIDGET, DRAWING TOOLS, WMS LAYERS #
+        #########################################
+
+        # Create drawing tools
+        desired_drawtools = ['rectangle']
+        draw_control = deawidgets.create_drawcontrol(desired_drawtools)
+
+        # Begin by displaying an empty layer group, and update the group with desired WMS on interaction.
+        self.map_layers = LayerGroup(layers=())
+        self.map_layers.name = 'Map Overlays'
+
+        # Create map widget
+        self.m = deawidgets.create_map(map_center=(-28, 135), zoom_level=4)
+        self.m.layout = make_box_layout()
+
+        # Add tools to map widget
+        self.m.add_control(draw_control)
+        self.m.add_layer(self.map_layers)
+
+        # Update all maps to starting defaults
+        update_map_layers(self)
+
+        ############################
+        # WIDGETS FOR APP CONTROLS #
+        ############################
+
+        # Create parameter widgets
+        dropdown_basemap = deawidgets.create_dropdown(self.basemap_list,
+                                                      self.basemap_list[0][1])
+        dropdown_dealayer = deawidgets.create_dropdown(self.dealayer_list,
+                                                       self.dealayer_list[0][1])
+        dropdown_output = deawidgets.create_dropdown(self.output_list,
+                                                     self.output_list[0][1])
+        date_picker = deawidgets.create_datepicker(
+            value=datetime.date(2021, 12, 31))
+        dropdown_styles = deawidgets.create_dropdown(self.styles_list,
+                                                     self.styles_list[0])
+        slider_abs = widgets.IntRangeSlider(value=[50, 3000],
+                                            min=0,
+                                            max=10000,
+                                            step=25,
+                                            description='',
+                                            layout={'width': '85%'})
+        run_button = create_expanded_button("Export imagery", "info")
+
+        # Expandable advanced section
+        text_resolution = widgets.FloatText(value=30,
+                                            description='',
+                                            layout={'width': '100%',
+                                                    'margin': '0px',
+                                                    'padding': '0px'})
+        checkbox_pansharpen = deawidgets.create_checkbox(
+            self.pansharpen, 'Pansharpen Landsat')
+        slider_power = widgets.FloatSlider(value=1.0,
+                                           min=0.01,
+                                           max=1.0,
+                                           step=0.01,
+                                           description='',
+                                            layout={'width': '85%'})
+        checkbox_unsharp_mask = deawidgets.create_checkbox(
+            self.unsharp_mask, 'Enable', layout={'width': '100%'})
+        text_unsharp_mask_radius = widgets.FloatText(value=20,
+                                                     description='Radius',
+                                                     layout={'width': '100%',
+                                                             'margin': '0px',
+                                                             'padding': '0px',
+                                                             'display': 'none'})
+        text_unsharp_mask_amount = widgets.FloatText(value=0.3,
+                                                     description='Amount',
+                                                     layout={'width': '100%',
+                                                             'margin': '0px',
+                                                             'padding': '0px',
+                                                             'display': 'none'})
+        checkbox_max_size = deawidgets.create_checkbox(
+            self.unsharp_mask, 'Enable')
+        expand_box = widgets.VBox([HTML('Resolution (metres):'),
+                                   text_resolution,
+                                   checkbox_pansharpen,
+                                   HTML('</br>Apply power transformation to darken bright features:'),
+                                   slider_power,
+                                   HTML('</br>Apply unsharp masking to sharpen image:'),
+                                   checkbox_unsharp_mask,
+                                   text_unsharp_mask_radius,
+                                   text_unsharp_mask_amount,
+                                   HTML('</br>Override maximum size limit: (use with caution; may cause memory issues/crashes)'),
+                                   checkbox_max_size],
+                                  layout={'overflow': 'hidden'},
+                                 )
+
+        expand = widgets.Accordion(children=[expand_box], selected_index=None)
+        expand.set_title(0, 'Advanced')
+
+        # Add specific dialogs to class so they can be modified
+        self.text_resolution = text_resolution
+        self.checkbox_pansharpen = checkbox_pansharpen
+        self.text_unsharp_mask_radius = text_unsharp_mask_radius
+        self.text_unsharp_mask_amount = text_unsharp_mask_amount
+
+        ####################################
+        # UPDATE FUNCTIONS FOR EACH WIDGET #
+        ####################################
+
+        # Run update functions whenever various widgets are changed.
+        date_picker.observe(self.update_date, "value")
+        dropdown_basemap.observe(self.update_basemap, "value")
+        dropdown_dealayer.observe(self.update_dealayer, "value")
+        dropdown_styles.observe(self.update_styles, "value")
+        dropdown_output.observe(self.update_output, "value")
+        run_button.on_click(self.run_app)
+        draw_control.on_draw(update_geojson)
+        slider_abs.observe(self.update_slider_abs, "value")
+        
+        # Advanced params
+        text_resolution.observe(self.update_text_resolution, "value")
+        checkbox_pansharpen.observe(self.update_checkbox_pansharpen, "value")
+        slider_power.observe(self.update_slider_power, "value")
+        checkbox_unsharp_mask.observe(self.update_checkbox_unsharp_mask, "value")
+        text_unsharp_mask_radius.observe(self.update_text_unsharp_mask_radius, "value")
+        text_unsharp_mask_amount.observe(self.update_text_unsharp_mask_amount, "value")
+        checkbox_max_size.observe(self.update_checkbox_max_size, "value")
+        
+
+        ##################################
+        # COLLECTION OF ALL APP CONTROLS #
+        ##################################
+
+        parameter_selection = VBox([
+            HTML("<b>Date:</b>"),
+            date_picker,
+            HTML('<b>Satellite imagery:</b>'),
+            dropdown_dealayer,
+            HTML('<b>Style:</b>'),
+            dropdown_styles,
+            HTML('<b>Colour stretch:</b>'),
+            slider_abs,
+            HTML('<b>Output file format:</b>'),
+            dropdown_output,
+            HTML('</br>'),
+            expand,
+        ])
+        map_selection = VBox([
+            HTML("</br><b>Map overlay:</b>"),
+            dropdown_basemap,
+        ])
+        parameter_selection.layout = make_box_layout()
+        map_selection.layout = make_box_layout()
+
+        ###############################
+        # SPECIFICATION OF APP LAYOUT #
+        ###############################
+
+        #       0   1    2   3   4   5   6   7    8   9
+        #     ---------------------------------------------
+        # 0   | Header                         | Map sel. |
+        #     ---------------------------------------------
+        # 1   | Params |                                  |
+        # 2   |        |                                  |
+        # 3   |        |                                  |
+        # 4   |        |               Map                |
+        # 5   |        |                                  |
+        #     ----------                                  |
+        # 6   |  Run   |                                  |
+        #     ---------------------------------------------
+        # 7   |   Status info   |      Figure/output      |
+        # 8   |                 |                         |
+        # 9   |                 |                         |
+        # 10  |                 |                         |
+        # 11  | ------------------------------------------|
+
+        # Create the layout #[rowspan, colspan]
+        grid = GridspecLayout(12, 10, height="1400px", width="auto")
+
+        # Header and controls
+        grid[0, :8] = self.header
+        grid[0, 8:] = map_selection
+        grid[1:6, 0:2] = parameter_selection
+        grid[6, 0:2] = run_button
+
+        # Status info, map and plot
+        grid[1:7, 2:] = self.m  # map
+        grid[7:, 0:4] = self.status_info
+        grid[7:, 4:] = self.output_plot
+
+        # Display using HBox children attribute
+        self.children = [grid]
+
+    ######################################
+    # DEFINITION OF ALL UPDATE FUNCTIONS #
+    ######################################
+
+    # Update date
+    def update_date(self, change):
+        self.date = str(change.new)
+        update_map_layers(self)
+
+    # Update colour stretch
+    def update_slider_abs(self, change):
+        self.vmin, self.vmax = change.new
+    
+    # Update power transform
+    def update_slider_power(self, change):
+        self.power = change.new
+
+    # Enable pansharpening and reset/deactivate resolution
+    def update_checkbox_pansharpen(self, change):
+        self.pansharpen = change.new
+
+        # Override default resolution if pansharpening is specified;
+        # disable input if so
+        if change.new:
+            self.text_resolution.value = 15
+            self.text_resolution.disabled = True
+        else:
+            self.text_resolution.value = 30
+            self.text_resolution.disabled = False
+            
+    # Enable unsharp masking and show/hide custom params
+    def update_checkbox_unsharp_mask(self, change):
+        self.unsharp_mask = change.new
+        
+        # Show unsharp masking params in menu if activated
+        if change.new:
+            self.text_unsharp_mask_radius.layout.display = 'block'
+            self.text_unsharp_mask_amount.layout.display = 'block'
+        else:
+            self.text_unsharp_mask_radius.layout.display = 'none'
+            self.text_unsharp_mask_amount.layout.display = 'none'
+            
+    # Change unsharp masking radius
+    def update_text_unsharp_mask_radius(self, change):
+        self.unsharp_mask_radius = change.new
+        
+    # Change unsharp masking amount
+    def update_text_unsharp_mask_amount(self, change):
+        self.unsharp_mask_amount = change.new
+        
+    # Override max size limit
+    def update_checkbox_max_size(self, change):
+        self.max_size = change.new
+
+    # Update resolution
+    def update_text_resolution(self, change):
+        self.resolution = change.new
+
+        # Clear data load params to trigger data re-load
+        self.rgb_array = None
+        self.sensor = None
+        self.load_params = None
+        self.query_params = None
+
+    # Change layers shown on the map
+    def update_dealayer(self, change):
+        self.dealayer = change.new
+
+        if change.new == 'ga_ls_ard_3':
+            self.text_resolution.value = 30
+            self.checkbox_pansharpen.disabled = False
+
+            if self.pansharpen:
+                self.text_resolution.value = 15
+                self.text_resolution.disabled = True
+
+        else:
+            self.text_resolution.value = 10
+            self.checkbox_pansharpen.disabled = True
+            self.text_resolution.disabled = False
+
+        update_map_layers(self)
+
+    # Update basemap
+    def update_basemap(self, change):
+        self.basemap = change.new
+        update_map_layers(self)
+
+    # Set imagery style
+    def update_styles(self, change):
+        self.style = change.new
+        update_map_layers(self)
+
+    # Set output file format
+    def update_output(self, change):
+        self.output_format = change.new
+
+    def run_app(self, change):
+
+        # Clear progress bar and output areas before running
+        self.status_info.clear_output()
+        self.output_plot.clear_output()
+        
+        # Verify that polygon was drawn
+        if self.gdf_drawn is not None:
+
+            with self.status_info:
+
+                    # Load data and add to attribute
+                    if self.rgb_array is None:
+                        self.rgb_array = extract_data(self)                            
+                    else:
+                        print('Using previously loaded data')                        
+
+            if self.rgb_array is not None:
+                
+                with self.status_info:
+
+                    # Create unique file name
+                    centre_coords = self.gdf_drawn.geometry[0].centroid.coords[0][::-1]
+                    site = reverse_geocode(coords=centre_coords)
+                    fname = (f"{self.sensor} - {self.date} - {site} - {self.style}, "
+                             f"{self.resolution:.0f} m resolution.{self.output_format}")
+
+                    # Remove spaces and commas if requested
+                    if self.standardise_name:
+                        fname = (fname
+                                 .replace(' - ', '_')
+                                 .replace(', ', '-')
+                                 .replace(' ', '-').lower())
+
+                    print(f'\nExporting image for {site}.\nThis may take several minutes...')
+
+                ############
+                # Plotting #
+                ############
+
+                with self.output_plot:            
+                    plot_data(self, fname)
+            
+            else:
+                with self.status_info:
+                    print('No satellite data found in the selected area. '
+                          'Please select a new rectangle over an area with '
+                          'satellite imagery visible on the map.')
+        
+        else:
+            with self.status_info:
+                print('Please draw a rectangle on the map, then press "Export imagery"')
+
